[CORRECTIVE] Updated the addition of address block graphics items to memory map visualization
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@620 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ IPXactWrapper/ComponentEditor/memoryMaps/memoryMapsVisualizer/registergraphitem.cpp
@@ -103,7 +103,7 @@ void RegisterGraphItem::reorganizeChildren() {
 	qreal xCoordinate = rect().left();
 	MemoryVisualizationItem* previous = NULL;
 
-	for (QMap<int, MemoryVisualizationItem*>::iterator i = --childItems_.end();
+	for (QMap<quint64, MemoryVisualizationItem*>::iterator i = --childItems_.end();
 		i != --childItems_.begin(); --i) {
 			
 			Q_ASSERT(i.value());
